Extract AgentExportMeterSupplier for agent metric export path
Replace anonymous MeterSupplier in APMMeterService with a named type.

Made-with: Cursor
